Move render hook helper to `wrappedTestingLibrary/hooks`.
