+++ graylog2-web-interface/src/hooks/useFavoriteItemMutation.test.ts
@@ -14,7 +14,7 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-import { renderHook, renderHookAct } from 'wrappedTestingLibrary';
+import { renderHook, act } from 'wrappedTestingLibrary/hooks';
 
 import asMock from 'helpers/mocking/AsMock';
 import fetch from 'logic/rest/FetchProvider';
@@ -49,7 +49,7 @@ describe('useFavoriteItemMutation', () => {
       asMock(fetch).mockImplementation(() => Promise.resolve({}));
       const { result, waitFor } = renderHook(() => useUserSearchFilterMutation(), { queryClientOptions: { logger } });
 
-      renderHookAct(() => {
+      act(() => {
         result.current.putItem('111');
       });
 
@@ -62,7 +62,7 @@ describe('useFavoriteItemMutation', () => {
 
       const { result, waitFor } = renderHook(() => useUserSearchFilterMutation(), { queryClientOptions: { logger } });
 
-      renderHookAct(() => {
+      act(() => {
         result.current.putItem('111').catch(() => {});
       });
 
@@ -79,7 +79,7 @@ describe('useFavoriteItemMutation', () => {
       asMock(fetch).mockImplementation(() => Promise.resolve());
       const { result, waitFor } = renderHook(() => useUserSearchFilterMutation(), { queryClientOptions: { logger } });
 
-      renderHookAct(() => {
+      act(() => {
         result.current.deleteItem('111');
       });
 
@@ -91,7 +91,7 @@ describe('useFavoriteItemMutation', () => {
 
       const { result, waitFor } = renderHook(() => useUserSearchFilterMutation(), { queryClientOptions: { logger } });
 
-      renderHookAct(() => {
+      act(() => {
         result.current.deleteItem('111').catch(() => {});
       });
