+++ graylog2-web-interface/test/wrappedTestingLibrary/index.tsx
@@ -14,43 +14,19 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-import * as React from 'react';
+import type * as React from 'react';
 import type { RenderOptions } from '@testing-library/react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import '@testing-library/jest-dom';
 import 'jest-styled-components';
-import type { RenderHookResult, RenderHookOptions } from '@testing-library/react-hooks';
-import { renderHook, act as renderHookAct } from '@testing-library/react-hooks';
-import type { QueryClientConfig } from '@tanstack/react-query';
 
-import DefaultQueryClientProvider from '../DefaultQueryClientProvider';
 import WrappingContainer from '../WrappingContainer';
 
 export const renderWithWrapper = (Component: React.ReactElement<any>, options?: RenderOptions) => render(Component, {
   wrapper: WrappingContainer,
   ...options,
 });
-
-const renderHookWithWrapper = <TProps, TResult>(
-  callback: (props: TProps) => TResult,
-  options: RenderHookOptions<TProps> & { queryClientOptions?: QueryClientConfig } = {},
-): RenderHookResult<TProps, TResult> => renderHook(
-    callback,
-    {
-      ...options,
-      wrapper: ({ children }) => {
-        const CustomWrapper = options.wrapper as React.ElementType ?? React.Fragment;
-
-        return (
-          <DefaultQueryClientProvider options={options.queryClientOptions}>
-            <CustomWrapper>
-              {children}
-            </CustomWrapper>
-          </DefaultQueryClientProvider>
-        );
-      },
-    });
 
 export function asElement<T extends new(...args: any) => any> (elem: any, elementType: T): InstanceType<T> {
   if (elem && elem instanceof elementType) {
@@ -64,7 +40,5 @@ export function asElement<T extends new(...args: any) => any> (elem: any, elemen
 
 export * from '@testing-library/react';
 export {
-  renderHookAct,
   renderWithWrapper as render,
-  renderHookWithWrapper as renderHook,
 };
